wrap CRUD methods in if-else statements to check that the current_user id matches what was entered, otherwise redirect

+++ app/controllers/materials_controller.rb
@@ -5,62 +5,74 @@ class MaterialsController < ApplicationController
 
     matching_materials = Material.where({ :id => the_id })
 
-    @the_material = matching_materials.at(0)
+    if matching_materials.at(0).user_id == current_user.id
+      @the_material = matching_materials.at(0)
 
-    render({ :template => "materials/show" })
+      render({ :template => "materials/show" })
+    else
+      redirect_to("/production_plans", {:notice => "You don't have access to that page."})
+    end
+
   end
 
   def create
     the_material = Material.new
-    the_material.description = params.fetch("query_description")
-    the_material.crop_time = params.fetch("query_crop_time")
-    the_material.container_type = params.fetch("query_container_type")
-    the_material.container_size = params.fetch("query_container_size")
-    the_material.total_qty = params.fetch("query_total_qty")
-
-    the_material.material_cost = params.fetch("query_material_cost")
-    the_material.unit_container_cost = params.fetch("query_unit_container_cost")
-    the_material.unit_tag_cost = params.fetch("query_unit_tag_cost")
-    the_material.miscellaneous = params.fetch("query_miscellaneous")
-    
-    the_material.buffer = params.fetch("query_buffer")
-    
-    the_material.unit_price = params.fetch("query_unit_price")
     
     the_material.user_id = current_user.id
     the_material.plan_id = params.fetch("query_plan_id").to_i
 
-    diam = the_material.container_size
-    
-    if the_material.container_type == "Pot"
-      the_material.bench_space = ((diam**2/144.0) * the_material.total_qty).round(2)
-    elsif the_material.container_type == "Tray"
-      the_material.bench_space = (((11*22)/144.0) * (the_material.total_qty / diam)).round(2)
-    end
+    check_prodplan = ProductionPlan.where({:id => the_material.plan_id})
 
-    the_material.total_bench_space_weeks = the_material.bench_space * the_material.crop_time
-    prodplan = ProductionPlan.where({:id=>params.fetch("query_plan_id").to_i}).at(0)
+    if check_prodplan.user_id == current_user.id
+      the_material.description = params.fetch("query_description")
+      the_material.crop_time = params.fetch("query_crop_time")
+      the_material.container_type = params.fetch("query_container_type")
+      the_material.container_size = params.fetch("query_container_size")
+      the_material.total_qty = params.fetch("query_total_qty")
 
-    if the_material.container_type == "Pot"
-      the_material.soil_cost = (((3.14 * ((diam/2)**2) * diam) / 46656.0) * prodplan.soil_cost).round(2)
-    elsif matl.container_type == "Tray"
-      cell_volume = Math.sqrt((11*22.0)/diam)**3
+      the_material.material_cost = params.fetch("query_material_cost")
+      the_material.unit_container_cost = params.fetch("query_unit_container_cost")
+      the_material.unit_tag_cost = params.fetch("query_unit_tag_cost")
+      the_material.miscellaneous = params.fetch("query_miscellaneous")
+      
+      the_material.buffer = params.fetch("query_buffer")
+      
+      the_material.unit_price = params.fetch("query_unit_price")
 
-      matl.soil_cost = ((cell_volume / 46656) * prodplan.soil_cost).round(2)
-    end
-    
-    the_material.shrink_opportunity_cost = ((the_material.buffer/100.0) * the_material.unit_price).round(2)
+      diam = the_material.container_size
+      
+      if the_material.container_type == "Pot"
+        the_material.bench_space = ((diam**2/144.0) * the_material.total_qty).round(2)
+      elsif the_material.container_type == "Tray"
+        the_material.bench_space = (((11*22)/144.0) * (the_material.total_qty / diam)).round(2)
+      end
 
-    the_material.unit_cost = the_material.shrink_opportunity_cost + the_material.unit_container_cost + the_material.unit_tag_cost + the_material.material_cost + the_material.soil_cost
+      the_material.total_bench_space_weeks = the_material.bench_space * the_material.crop_time
+      prodplan = ProductionPlan.where({:id=>params.fetch("query_plan_id").to_i}).at(0)
 
-    the_material.total_cost = the_material.unit_cost * the_material.total_qty
-    the_material.total_revenue = the_material.unit_price * the_material.total_qty
+      if the_material.container_type == "Pot"
+        the_material.soil_cost = (((3.14 * ((diam/2)**2) * diam) / 46656.0) * prodplan.soil_cost).round(2)
+      elsif matl.container_type == "Tray"
+        cell_volume = Math.sqrt((11*22.0)/diam)**3
 
-    if the_material.valid?
-      the_material.save
-      redirect_to("/production_plans/#{the_material.plan_id}", { :notice => "Material created successfully." })
+        matl.soil_cost = ((cell_volume / 46656) * prodplan.soil_cost).round(2)
+      end
+      
+      the_material.shrink_opportunity_cost = ((the_material.buffer/100.0) * the_material.unit_price).round(2)
+
+      the_material.unit_cost = the_material.shrink_opportunity_cost + the_material.unit_container_cost + the_material.unit_tag_cost + the_material.material_cost + the_material.soil_cost
+
+      the_material.total_cost = the_material.unit_cost * the_material.total_qty
+      the_material.total_revenue = the_material.unit_price * the_material.total_qty
+
+      if the_material.valid?
+        the_material.save
+        redirect_to("/production_plans/#{the_material.plan_id}", { :notice => "Material created successfully." })
+      else
+        redirect_to("/production_plans/#{the_material.plan_id}", { :alert => the_material.errors.full_messages.to_sentence })
+      end
     else
-      redirect_to("/production_plans/#{the_material.plan_id}", { :alert => the_material.errors.full_messages.to_sentence })
+      redirect_to("/production_plans")
     end
   end
 
@@ -68,14 +80,19 @@ class MaterialsController < ApplicationController
     the_id = params.fetch("path_id")
     the_material = Material.where({ :id => the_id }).at(0)
 
-    the_material2 = the_material.dup
+    if the_material.user_id == current_user.id
 
-    if the_material2.valid?
-      the_material2.save
-      
-      redirect_to("/production_plans/#{the_material.plan_id}", { :notice => "Material copied successfully." })
+      the_material2 = the_material.dup
+
+      if the_material2.valid?
+        the_material2.save
+        
+        redirect_to("/production_plans/#{the_material.plan_id}", { :notice => "Material copied successfully." })
+      else
+        redirect_to("/production_plans/#{the_material.plan_id}", { :alert => the_material.errors.full_messages.to_sentence })
+      end
     else
-      redirect_to("/production_plans/#{the_material.plan_id}", { :alert => the_material.errors.full_messages.to_sentence })
+      redirect_to("/production_plans", {:notice => "You don't have access to that page."})
     end
 
   end
@@ -83,59 +100,69 @@ class MaterialsController < ApplicationController
   def update
     the_id = params.fetch("path_id")
     the_material = Material.where({ :id => the_id }).at(0)
-    the_material.description = params.fetch("query_description")
-    the_material.crop_time = params.fetch("query_crop_time")
-    the_material.container_type = params.fetch("query_container_type")
-    the_material.container_size = params.fetch("query_container_size")
-    the_material.total_qty = params.fetch("query_total_qty")
 
-    the_material.material_cost = params.fetch("query_material_cost")
-    the_material.unit_container_cost = params.fetch("query_unit_container_cost")
-    the_material.unit_tag_cost = params.fetch("query_unit_tag_cost")
-    the_material.miscellaneous = params.fetch("query_miscellaneous")
+    if the_material.user_id == current_user.id
+      the_material.description = params.fetch("query_description")
+      the_material.crop_time = params.fetch("query_crop_time")
+      the_material.container_type = params.fetch("query_container_type")
+      the_material.container_size = params.fetch("query_container_size")
+      the_material.total_qty = params.fetch("query_total_qty")
+
+      the_material.material_cost = params.fetch("query_material_cost")
+      the_material.unit_container_cost = params.fetch("query_unit_container_cost")
+      the_material.unit_tag_cost = params.fetch("query_unit_tag_cost")
+      the_material.miscellaneous = params.fetch("query_miscellaneous")
+      
+      the_material.buffer = params.fetch("query_buffer")
+      
+      the_material.unit_price = params.fetch("query_unit_price")
+      
+      the_material.user_id = current_user.id
+      the_material.plan_id = the_material.plan_id
+
+      diam = the_material.container_size
+      
+      the_material.bench_space = ((diam**2/144.0) * the_material.total_qty).round(2)
+      the_material.total_bench_space_weeks = the_material.bench_space * the_material.crop_time
+      prodplan = ProductionPlan.where({:id=>the_material.plan_id}).at(0)
+
+      if the_material.container_type == "Pot"
+        the_material.soil_cost = (((3.14 * ((diam/2)**2) * diam) / 46656.0) * prodplan.soil_cost).round(2)
+      elsif the_material.container_type == "Tray"
+        cell_volume = Math.sqrt((11*22.0)/diam)**3
+
+        the_material.soil_cost = ((cell_volume / 46656) * prodplan.soil_cost).round(2)
+      end
+      
+      the_material.shrink_opportunity_cost = ((the_material.buffer/100.0) * the_material.unit_price).round(2)
+
+      the_material.unit_cost = the_material.shrink_opportunity_cost + the_material.unit_container_cost + the_material.unit_tag_cost + the_material.material_cost + the_material.soil_cost
+
+      the_material.total_cost = the_material.unit_cost * the_material.total_qty
+      the_material.total_revenue = the_material.unit_price * the_material.total_qty
+
+      if the_material.valid?
+        the_material.save
+        redirect_to("/production_plans/#{the_material.plan_id}", { :notice => "Material updated successfully."} )
+      else
+        redirect_to("/materials/#{the_material.id}", { :alert => the_material.errors.full_messages.to_sentence })
+      end
     
-    the_material.buffer = params.fetch("query_buffer")
-    
-    the_material.unit_price = params.fetch("query_unit_price")
-    
-    the_material.user_id = current_user.id
-    the_material.plan_id = the_material.plan_id
-
-    diam = the_material.container_size
-    
-    the_material.bench_space = ((diam**2/144.0) * the_material.total_qty).round(2)
-    the_material.total_bench_space_weeks = the_material.bench_space * the_material.crop_time
-    prodplan = ProductionPlan.where({:id=>the_material.plan_id}).at(0)
-
-    if the_material.container_type == "Pot"
-      the_material.soil_cost = (((3.14 * ((diam/2)**2) * diam) / 46656.0) * prodplan.soil_cost).round(2)
-    elsif the_material.container_type == "Tray"
-      cell_volume = Math.sqrt((11*22.0)/diam)**3
-
-      the_material.soil_cost = ((cell_volume / 46656) * prodplan.soil_cost).round(2)
-    end
-    
-    the_material.shrink_opportunity_cost = ((the_material.buffer/100.0) * the_material.unit_price).round(2)
-
-    the_material.unit_cost = the_material.shrink_opportunity_cost + the_material.unit_container_cost + the_material.unit_tag_cost + the_material.material_cost + the_material.soil_cost
-
-    the_material.total_cost = the_material.unit_cost * the_material.total_qty
-    the_material.total_revenue = the_material.unit_price * the_material.total_qty
-
-    if the_material.valid?
-      the_material.save
-      redirect_to("/production_plans/#{the_material.plan_id}", { :notice => "Material updated successfully."} )
     else
-      redirect_to("/materials/#{the_material.id}", { :alert => the_material.errors.full_messages.to_sentence })
+      redirect_to("/production_plans", {:notice => "You don't have access to that page."})
     end
   end
 
   def destroy
     the_id = params.fetch("path_id")
     the_material = Material.where({ :id => the_id }).at(0)
+    
+    if the_material.user_id == current_user.id
+      the_material.destroy
 
-    the_material.destroy
-
-    redirect_to("/production_plans/#{the_material.plan_id}", { :notice => "Material deleted successfully."} )
+      redirect_to("/production_plans/#{the_material.plan_id}", { :notice => "Material deleted successfully."} )
+    else
+      redirect_to("/production_plans", {:notice => "You don't have access to that page."})
+    end
   end
 end
